Optimization of JIT buffer relocation
When allocating JIT bufer close to PHP .text segment, here we modify the
code to avoid the problem that the algorithm fails to find a suitable
candidate because only the first half of the free address segment is verified.

Before PHP .text segment, search and verify the second half of each free segment;
after PHP .text segment, verify the first half of each free segment, which avoids
missing possible candidates due to the large address range of the free segment.

Fixs issue #####.

Signed-off-by: Long, Tao <tao.long@intel.com>